requirements.txt
alembic==1.14.1
annotated-types==0.7.0
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
cachelib==0.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
eventlet==0.39.0
firebase==4.0.1
flasgger==0.9.7.1
Flask==3.1.0
Flask-Caching==2.3.0
Flask-JSON==0.4.0
Flask-JWT==0.3.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-mongoengine==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
MarkupSafe==3.0.2
mistune==3.1.0
mongoengine==0.29.1
outcome==1.3.0.post0
packaging==24.2
passkeys==2.0.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.10.1
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
schedule==1.2.2
selenium==4.28.1
sentry-sdk==2.20.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
waitress==3.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
